requirements.txt
numpy==1.22.3
pandas==1.4.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
debugpy==1.6.0
decorator==5.1.1
distro==1.7.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.4
executing==0.8.3
greenlet==1.1.2
idna==3.3
ipykernel==6.13.0
ipython==8.2.0
jedi==0.18.1
jsonschema==3.2.0
jupyter-client==7.2.2
jupyter-core==4.10.0
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.22.3
packaging==21.3
pandas==1.4.2
paramiko==2.10.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
pymongo==4.1.1
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.35
stack-data==0.2.0
texttable==1.6.4
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
urllib3==1.26.9
wcwidth==0.2.5
websocket-client==0.59.0
